meshの3D表示

+++ src/gl_renderer/mesh.rs
@@ -1,15 +1,16 @@
 use glam::Vec3;
 
+#[derive(Debug)]
 pub struct Mesh {
     pub vertices: Vec<MeshVertex>,
     pub triangles: Vec<[usize; 3]>,
 }
 
 impl Mesh {
-    pub fn get_flat_vertices(&self) -> (Vec<f32>, Vec<u8>, Vec<f32>) {
+    pub fn get_flat_vertices(&self) -> (Vec<f32>, Vec<f32>, Vec<f32>) {
         let vertex_count = self.triangles.len() * 3;
         let mut positions: Vec<f32> = Vec::with_capacity(vertex_count * 3);
-        let mut colors: Vec<u8> = Vec::with_capacity(vertex_count * 4);
+        let mut colors: Vec<f32> = Vec::with_capacity(vertex_count * 4);
         let mut normals: Vec<f32> = Vec::with_capacity(vertex_count * 3);
 
         for indices in &self.triangles {
@@ -25,21 +26,23 @@ impl Mesh {
     }
 }
 
+#[derive(Debug)]
 pub struct MeshVertex {
     pub position: Vec3,
     pub color: Color4,
     pub normal: Vec3,
 }
 
+#[derive(Debug)]
 pub struct Color4 {
-    pub r: u8,
-    pub g: u8,
-    pub b: u8,
-    pub a: u8,
+    pub r: f32,
+    pub g: f32,
+    pub b: f32,
+    pub a: f32,
 }
 
 impl Color4 {
-    pub fn to_array(&self) -> [u8; 4] {
+    pub fn to_array(&self) -> [f32; 4] {
         [self.r, self.g, self.b, self.a]
     }
 }
